trainforge/cli/requirements.txt
# File: trainforge/cli/requirements.txt
# CLI dependencies for TrainForge

click>=8.0.0          # Command-line interface framework
requests>=2.28.0      # HTTP client for API calls
pyyaml>=6.0           # YAML configuration parsing
colorama>=0.4.4       # Colored terminal output
tabulate>=0.9.0       # Pretty table formatting